refactor: extracted a couple of functions
