+++ packages/gatsby-theme-docs/src/utils/navigate-with-filters.js
@@ -1,0 +1,21 @@
+import { navigate } from '@reach/router';
+import extractQueryParameters from './extract-query-parameters';
+
+function navigateWithFilters(filters, location) {
+  const currentQueryParameters = extractQueryParameters(location);
+  const newQueryParameters = { ...currentQueryParameters, ...filters };
+
+  newQueryParameters.filterTopics = JSON.stringify(
+    newQueryParameters.filterTopics
+  );
+
+  navigate(
+    `?fromFilterDate=${encodeURIComponent(
+      newQueryParameters.fromFilterDate
+    )}&toFilterDate=${encodeURIComponent(
+      newQueryParameters.toFilterDate
+    )}&filterTopics=${encodeURIComponent(newQueryParameters.filterTopics)}`
+  );
+}
+
+export default navigateWithFilters;
